adding 64-bit support (wip)

+++ src/determiniser/common/x86_common.h
@@ -4,7 +4,7 @@
 #include "remote_loader.h"
 #include <stdint.h>
 
-void x86_trap_handler (uint32_t * gregs, uint32_t trapno);
+void x86_trap_handler (uintptr_t * gregs, uint32_t trapno);
 void x86_startup (size_t minPage, size_t maxPage, CommStruct * pcs);
 void x86_check_version (CommStruct * pcs);
 void x86_bp_trap (int code, void * arg);
